+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   - Removes the `doctrine.config.metadata.reader` service definition
   - Adds `ExplicitClasslistAttributeDriver` as the default mapping driver
   - Drops composer dependency on doctrine/annotations
+* [BREAKING] Switch to using PSR-6 cache implementations for metadata, query and result caches:
+  - drops the composer dependency on doctrine/cache in favour of symfony/cache
+  - DoctrineCacheFactory methods are now hard-typehinted to return CacheItemPoolInterface
 * Drop support for PHP < 8.2
 
 ## v2.1.0 (2024-09-25)